Krushi bazar/requirements.txt
Flask==2.3.3
PyMySQL==1.1.0
python-dotenv==1.0.0
requests==2.31.0
APScheduler==3.10.4
aiohttp==3.9.1
